Move Assist Pipeline tests to right file
